apps/auth_service/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = "Runote authentication microservice"
authors = ["Syuhei Fujioka"]
readme = "feature.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pydantic = "^2.6.0"
sqlalchemy = "^2.0.28"
mysql-connector-python = "^8.3.0"
python-dotenv = "^1.0.1"
pymysql = "^1.1.1"
cryptography = "^44.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
httpx = "^0.27.0"
ruff = "^0.3.0"
pytest-asyncio = "^0.23.5"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"